feat: upgrade raylib

See merge request gtor/niseci-project/esox!39

+++ src/main.rs
@@ -179,14 +179,14 @@ fn main() {
     }
 
     // 10 is way too small for the default font height
-    let gui_default_font_height: i32 = rl.gui_get_style(DEFAULT, TEXT_SIZE as i32) *2;
-    rl.gui_set_style(DEFAULT, TEXT_SIZE as i32, gui_default_font_height);
+    let gui_default_font_height: i32 = rl.gui_get_style(DEFAULT, TEXT_SIZE) *2;
+    rl.gui_set_style(DEFAULT, TEXT_SIZE, gui_default_font_height);
     let gui_current_font_height: i32 = gui_default_font_height;
 
-    let txt_color_int = rl.gui_get_style(DEFAULT, TEXT_COLOR_NORMAL as i32);
-    let bg_color_int = rl.gui_get_style(DEFAULT, BACKGROUND_COLOR as i32);
-    let txt_spacing = rl.gui_get_style(DEFAULT, TEXT_SPACING as i32) *2;
-    rl.gui_set_style(DEFAULT, TEXT_SPACING as i32, txt_spacing);
+    let txt_color_int = rl.gui_get_style(DEFAULT, TEXT_COLOR_NORMAL);
+    let bg_color_int = rl.gui_get_style(DEFAULT, BACKGROUND_COLOR);
+    let txt_spacing = rl.gui_get_style(DEFAULT, TEXT_SPACING) *2;
+    rl.gui_set_style(DEFAULT, TEXT_SPACING, txt_spacing);
     let current_font = rl.gui_get_font();
     let mut main_state = MainState::new(
         gui_default_font_height,
